requirements.txt
amqp==2.2.2
anyjson==0.3.3
billiard==3.5.0.3
celery==4.1.0
Django==1.11.6
django-celery==3.2.1
django-redis-sessions==0.6.1
django-tinymce==2.6.0
kombu==4.1.0
mysqlclient==1.3.12
olefile==0.44
Pillow==4.3.0
pytz==2017.2
redis==2.10.6
vine==1.1.4
